[vim] .vimrc: Wrote SetupDictionary, au'd it
Based almost purely on the work of "Matt Wozniski-2" on the vim mailing
list.

I expanded slightly on Matt's original code by adding the handlling of
prefixes and such.

I should probably check to see somehow if this has a major affect on
vim's performance when starting, and if so work out how to optimise it,
but eh. 'sall good.